EDMT-414 refactor : 학생 1명의 생활기록부 정보를 가져오는 api 명세 변경에 따른 코드 리팩토링 및 필요 없는 파일 제거
